# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SWAN GMAO API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assets:
    get:
      operationId: getAssets
      tags: [assets]
      summary: List all assets
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Create a new asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssetBody"
      responses:
        "201":
          description: Created asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{id}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Get asset by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Asset details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    put:
      operationId: updateAsset
      tags: [assets]
      summary: Update an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssetBody"
      responses:
        "200":
          description: Updated asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    delete:
      operationId: deleteAsset
      tags: [assets]
      summary: Delete an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /assets/{id}/workorders:
    get:
      operationId: getAssetWorkOrders
      tags: [assets]
      summary: Get work orders for an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Work orders for asset
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkOrder"
  /assets/{id}/parts:
    get:
      operationId: getAssetParts
      tags: [assets]
      summary: Get spare parts catalog for an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Asset parts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssetPart"
    post:
      operationId: createAssetPart
      tags: [assets]
      summary: Add a part to the asset parts catalog
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssetPartBody"
      responses:
        "201":
          description: Created asset part
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetPart"
  /assets/{id}/parts/{partId}:
    put:
      operationId: updateAssetPart
      tags: [assets]
      summary: Update an asset part
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: partId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssetPartBody"
      responses:
        "200":
          description: Updated asset part
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetPart"
    delete:
      operationId: deleteAssetPart
      tags: [assets]
      summary: Remove a part from the asset parts catalog
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: partId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /sites:
    get:
      operationId: getSites
      tags: [sites]
      summary: List all sites
      responses:
        "200":
          description: List of sites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Site"
    post:
      operationId: createSite
      tags: [sites]
      summary: Create a site
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSiteBody"
      responses:
        "201":
          description: Created site
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
  /sites/{id}:
    put:
      operationId: updateSite
      tags: [sites]
      summary: Update a site
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSiteBody"
      responses:
        "200":
          description: Updated site
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
    delete:
      operationId: deleteSite
      tags: [sites]
      summary: Delete a site
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Asset:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        serialNumber:
          type: string
        location:
          type: string
        status:
          type: string
          enum: [operational, maintenance, breakdown, decommissioned]
        manufacturer:
          type: string
        model:
          type: string
        installDate:
          type: string
        lastMaintenanceDate:
          type: string
        mtbf:
          type: number
        mttr:
          type: number
        availabilityRate:
          type: number
        criticality:
          type: string
          enum: [low, medium, high, critical]
        siteId:
          type: integer
        siteName:
          type: string
        zoneId:
          type: integer
        zoneName:
          type: string
        photoUrl:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - category
        - status
        - criticality
        - createdAt
    CreateAssetBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        serialNumber:
          type: string
        location:
          type: string
        status:
          type: string
          enum: [operational, maintenance, breakdown, decommissioned]
        manufacturer:
          type: string
        model:
          type: string
        installDate:
          type: string
        criticality:
          type: string
          enum: [low, medium, high, critical]
        siteId:
          type: integer
        zoneId:
          type: integer
        photoUrl:
          type: string
      required:
        - name
        - category
        - status
        - criticality
    AssetPart:
      type: object
      properties:
        id:
          type: integer
        assetId:
          type: integer
        inventoryItemId:
          type: integer
        partName:
          type: string
        reference:
          type: string
        quantity:
          type: number
        unit:
          type: string
        note:
          type: string
        createdAt:
          type: string
        inventoryItemName:
          type: string
        inventoryItemQuantity:
          type: integer
        inventoryItemUnit:
          type: string
      required:
        - id
        - assetId
        - partName
        - quantity
        - createdAt
    CreateAssetPartBody:
      type: object
      properties:
        inventoryItemId:
          type: integer
        partName:
          type: string
        reference:
          type: string
        quantity:
          type: number
        unit:
          type: string
        note:
          type: string
      required:
        - partName
    WorkOrder:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [corrective, preventive, predictive, inspection]
        priority:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [open, in_progress, completed, cancelled, on_hold]
        assetId:
          type: integer
        assetName:
          type: string
        technicianId:
          type: integer
        technicianName:
          type: string
        estimatedHours:
          type: number
        actualHours:
          type: number
        scheduledDate:
          type: string
        completedDate:
          type: string
        siteId:
          type: integer
        siteName:
          type: string
        zoneId:
          type: integer
        zoneName:
          type: string
        assignmentMode:
          type: string
          enum: [by_technician, by_zone, by_machine, by_type]
        createdAt:
          type: string
      required:
        - id
        - title
        - type
        - priority
        - status
        - createdAt
...
    Site:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        city:
          type: string
        country:
          type: string
        zoneCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - country
        - createdAt
    CreateSiteBody:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        city:
          type: string
        country:
          type: string
      required:
        - name
...